+++ apps/storybook/src/playground/pages/rewards/affiliate/page.tsx
@@ -5,7 +5,7 @@ export default function AffiliatePage() {
     <ReferralProvider
       becomeAnAffiliateUrl="https://orderly.network/"
       learnAffiliateUrl="https://orderly.network/"
-      referralLinkUrl="/"
+      referralLinkUrl={window.location.origin}
       overwrite={{
         shortBrokerName: "Orderly",
         brokerName: "Orderly",