+++ packages/affiliate/src/pages/affiliate/referralLink/referralLink.script.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from "@orderly.network/i18n";
 import { toast } from "@orderly.network/ui";
 import { Decimal } from "@orderly.network/utils";
 import { useReferralContext } from "../../../provider";
-import { addQueryParam } from "../../../utils/utils";
+import { generateReferralLink } from "../../../utils/utils";
 
 export type ReferralLinkReturns = {
   onCopy?: (value: string) => void;
@@ -63,7 +63,7 @@ export const useReferralLinkScript = (): ReferralLinkReturns => {
   const referralLink = useMemo(() => {
     if (!firstCode) return "";
 
-    return addQueryParam(referralLinkUrl, "ref", firstCode.code);
+    return generateReferralLink(referralLinkUrl, firstCode.code);
   }, [firstCode]);
 
   const earn = useMemo(() => {