+++ packages/trading-leaderboard/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @orderly.network/trading-leaderboard
 
+## 2.8.4
+
+### Patch Changes
+
+- publish
+- 32b6043: publish
+- Updated dependencies
+- Updated dependencies [32b6043]
+  - @orderly.network/react-app@2.8.4
+  - @orderly.network/hooks@2.8.4
+  - @orderly.network/i18n@2.8.4
+  - @orderly.network/types@2.8.4
+  - @orderly.network/ui@2.8.4
+  - @orderly.network/ui-connector@2.8.4
+  - @orderly.network/utils@2.8.4
+
+## 2.8.4-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react-app@2.8.4-alpha.0
+  - @orderly.network/hooks@2.8.4-alpha.0
+  - @orderly.network/i18n@2.8.4-alpha.0
+  - @orderly.network/types@2.8.4-alpha.0
+  - @orderly.network/ui@2.8.4-alpha.0
+  - @orderly.network/ui-connector@2.8.4-alpha.0
+  - @orderly.network/utils@2.8.4-alpha.0
+
 ## 2.8.3
 
 ### Patch Changes